Move WeightInfo, Config, Event into Compliance Manager pallet.
